+++ configs/init.lua
@@ -1,13 +1,11 @@
 --- @type Awful
-local awful = require('awful')
+local awful = require("awful")
 
 --- @type Wibox
-local wibox = require('wibox')
+local wibox = require("wibox")
 
 --- @type Gears
-local gears = require('gears')
-
-
+local gears = require("gears")
 
 -- ██╗      █████╗ ██╗   ██╗ ██████╗ ██╗   ██╗████████╗███████╗
 -- ██║     ██╔══██╗╚██╗ ██╔╝██╔═══██╗██║   ██║╚══██╔══╝██╔════╝
@@ -18,10 +16,9 @@ local gears = require('gears')
 
 awful.layout.remove_default_layout()
 awful.layout.append_default_layouts({
-	awful.layout.suit.tile,
-	awful.layout.suit.floating,
+  awful.layout.suit.tile,
+  awful.layout.suit.floating,
 })
-
 
 -- ████████╗ █████╗  ██████╗ ███████╗
 -- ╚══██╔══╝██╔══██╗██╔════╝ ██╔════╝
@@ -31,14 +28,10 @@ awful.layout.append_default_layouts({
 --    ╚═╝   ╚═╝  ╚═╝ ╚═════╝ ╚══════╝
 
 screen.connect_signal("request::desktop_decoration", function(s)
-    awful.tag(
-            { "1", "2", "3", "4", "5", "6", "7", "8", "9", "0" },
-            s,
-            awful.layout.layouts[1]
-    )
+  awful.tag({ "1", "2", "3", "4", "5", "6", "7", "8", "9", "0" }, s, awful.layout.layouts[1])
 end)
 
--- ██╗    ██╗ █████╗ ██╗     ██╗     ██████╗  █████╗ ██████╗ ███████╗██████╗ 
+-- ██╗    ██╗ █████╗ ██╗     ██╗     ██████╗  █████╗ ██████╗ ███████╗██████╗
 -- ██║    ██║██╔══██╗██║     ██║     ██╔══██╗██╔══██╗██╔══██╗██╔════╝██╔══██╗
 -- ██║ █╗ ██║███████║██║     ██║     ██████╔╝███████║██████╔╝█████╗  ██████╔╝
 -- ██║███╗██║██╔══██║██║     ██║     ██╔═══╝ ██╔══██║██╔═══╝ ██╔══╝  ██╔══██╗
@@ -48,28 +41,30 @@ end)
 local wallpaper = gears.filesystem.get_random_file_from_dir(CONFIG.wallpaper_folder)
 -- Create a wibox for each screen and add it
 local function set_wallpaper(s)
-	-- Wallpaper
-	awful.wallpaper({
-		screen = s,
-		widget = {
-			image = CONFIG.wallpaper_folder .. "/" .. wallpaper,
-			resize = true,
-			horizontal_fit_policy = "fit",
-			vertical_fit_policy = "fit",
-			widget = wibox.widget.imagebox,
-		},
-	})
+  -- Wallpaper
+  awful.wallpaper({
+    screen = s,
+    widget = {
+      image = CONFIG.wallpaper_folder .. "/" .. wallpaper,
+      resize = true,
+      horizontal_fit_policy = "fit",
+      vertical_fit_policy = "fit",
+      widget = wibox.widget.imagebox,
+    },
+  })
 end
 
 -- Re-set wallpaper when a screen's geometry changes (e.g. different resolution)
 screen.connect_signal("property::geometry", set_wallpaper)
 
 awful.screen.connect_for_each_screen(function(s)
-	-- Wallpaper
-	set_wallpaper(s)
+  -- Wallpaper
+  set_wallpaper(s)
 end)
 
 ----------------------------------------
 -- Initialize basic configs
 ----------------------------------------
-load_all("configs", { "client", "Autostart", "keybindings"})
+require(... .. ".client")
+require(... .. ".Autostart")
+require(... .. ".keybindings")
