+++ error_handling.lua
@@ -1,0 +1,30 @@
+--- @type Naughty
+local naughty = require("naughty")
+
+-- Error handling during startup
+if awesome.startup_errors then
+  naughty.notify({
+    preset = naughty.config.presets.critical,
+    title = "Oops, there were errors during startup!",
+    text = awesome.startup_errors,
+  })
+end
+
+-- Handle runtime errors after startup
+do
+  local in_error = false
+  awesome.connect_signal("debug::error", function(err)
+    -- Make sure we don't go into an ondless error loop
+    if in_error then
+      return
+    end
+    in_error = true
+
+    naughty.notify({
+      preset = naughty.config.presets.critical,
+      title = "Oops, an error happened!",
+      text = tostring(err),
+    })
+    in_error = false
+  end)
+end
